chore: move component to where it is used
